自动超时优化

+++ lorne-tx-core/src/main/java/com/lorne/tx/db/LCNConnection.java
@@ -35,7 +35,7 @@ public class LCNConnection extends AbstractConnection {
             @Override
             public void run() {
                 System.out.println("自动回滚->" + getGroupId());
-                dataSourceService.schedule(getGroupId(),getCompensateList(), waitTask);
+                dataSourceService.schedule(getGroupId(), getCompensateList(), waitTask);
             }
         }, getMaxOutTime());
 
@@ -53,8 +53,11 @@ public class LCNConnection extends AbstractConnection {
         } else {
             connection.rollback();
         }
-        dataSourceService.deleteCompensates(getCompensateList());
+        if (rs != -100) {
+            dataSourceService.deleteCompensates(getCompensateList());
+        }
         waitTask.remove();
+
     }
